requirements.txt
opencv-python==4.0.0.21
Pillow==5.4.1
face-recognition==1.2.3
face-recognition-models==0.3.0
